+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## ⍺ Alpha
 
+#### 2025-01-29 — 0.0.2
+
+- On interface generation from a class, class implements interface and existing interfaces inherited by class are moved to the generated interface as inherited interfaces.
+- Added support for generics with Interface Implementations. Placeholder `%genericType%` is replaced with generic type.
+- Added additional default Interface Implementations.
+
 #### 2025-01-28 — 0.0.1
 
 - Initial release. Probably full of bugs. 🐛